+++ share/vss/pedersen/vss_test.go
@@ -123,7 +123,7 @@ func TestVSSShare(t *testing.T) {
 
 	aggr := ver.Aggregator
 
-	for i := 1; i < aggr.t-1; i++ {
+	for i := int64(1); i < aggr.t-1; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 	// not enough approvals
@@ -147,7 +147,7 @@ func TestVSSAggregatorDealCertified(t *testing.T) {
 	dealer := genDealer()
 	aggr := dealer.Aggregator
 
-	for i := 0; i < aggr.t; i++ {
+	for i := int64(0); i < aggr.t; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 
@@ -166,7 +166,7 @@ func TestVSSAggregatorDealCertified(t *testing.T) {
 
 	// inconsistent state on purpose
 	// too much complaints
-	for i := 0; i < aggr.t; i++ {
+	for i := int64(0); i < aggr.t; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusComplaint}
 	}
 	assert.False(t, aggr.DealCertified())
@@ -406,12 +406,12 @@ func TestVSSAggregatorAllResponses(t *testing.T) {
 	dealer := genDealer()
 	aggr := dealer.Aggregator
 
-	for i := 0; i < aggr.t; i++ {
+	for i := int64(0); i < aggr.t; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 	assert.False(t, aggr.DealCertified())
 
-	for i := aggr.t; i < nbVerifiers; i++ {
+	for i := aggr.t; i < int64(nbVerifiers); i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 
@@ -423,7 +423,7 @@ func TestVSSDealerTimeout(t *testing.T) {
 	dealer := genDealer()
 	aggr := dealer.Aggregator
 
-	for i := 0; i < aggr.t; i++ {
+	for i := int64(0); i < aggr.t; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 	require.False(t, aggr.DealCertified())
@@ -452,7 +452,7 @@ func TestVSSVerifierTimeout(t *testing.T) {
 	aggr := v.Aggregator
 
 	// Add t responses
-	for i := 0; i < aggr.t; i++ {
+	for i := int64(0); i < aggr.t; i++ {
 		aggr.responses[uint32(i)] = &Response{Status: StatusApproval}
 	}
 	assert.False(t, aggr.DealCertified())
@@ -534,16 +534,16 @@ func TestVSSAggregatorAddComplaint(t *testing.T) {
 func TestVSSSessionID(t *testing.T) {
 	dealer, _ := NewDealer(suite, dealerSec, secret, verifiersPub, vssThreshold)
 	commitments := dealer.deals[0].Commitments
-	sid, err := sessionID(suite, dealerPub, verifiersPub, commitments, dealer.t)
+	sid, err := sessionID(suite, dealerPub, verifiersPub, commitments, int(dealer.t))
 	assert.NoError(t, err)
 
-	sid2, err2 := sessionID(suite, dealerPub, verifiersPub, commitments, dealer.t)
+	sid2, err2 := sessionID(suite, dealerPub, verifiersPub, commitments, int(dealer.t))
 	assert.NoError(t, err2)
 	assert.Equal(t, sid, sid2)
 
 	wrongDealerPub := suite.Point().Add(dealerPub, dealerPub)
 
-	sid3, err3 := sessionID(suite, wrongDealerPub, verifiersPub, commitments, dealer.t)
+	sid3, err3 := sessionID(suite, wrongDealerPub, verifiersPub, commitments, int(dealer.t))
 	assert.NoError(t, err3)
 	assert.NotEqual(t, sid3, sid2)
 }
